cmd/fyne: Initial refactor of packaging code
