+++ cmd/fyne/package-darwin.go
@@ -1,0 +1,66 @@
+package main
+
+import (
+	"image"
+	"os"
+	"path/filepath"
+	"text/template"
+
+	"github.com/jackmordaunt/icns"
+	"github.com/pkg/errors"
+)
+
+type darwinData struct {
+	Name, ExeName string
+	AppID         string
+}
+
+func (p *packager) packageDarwin() error {
+	appDir := ensureSubDir(p.dir, p.name+".app")
+	exeName := filepath.Base(p.exe)
+
+	contentsDir := ensureSubDir(appDir, "Contents")
+	info := filepath.Join(contentsDir, "Info.plist")
+	infoFile, _ := os.Create(info)
+
+	tpl, err := template.ParseFiles("templates/Info.Plist")
+	if err != nil {
+		return errors.Wrap(err, "Failed to parse plist template")
+	}
+
+	tplData := darwinData{Name: p.name, ExeName: exeName, AppID: p.appID}
+	err = tpl.Execute(infoFile, tplData)
+	if err != nil {
+		return errors.Wrap(err, "Failed to write plist template")
+	}
+
+	macOSDir := ensureSubDir(contentsDir, "MacOS")
+	binName := filepath.Join(macOSDir, exeName)
+	err = copyExeFile(p.exe, binName)
+	if err != nil {
+		return errors.Wrap(err, "Failed to copy exe file")
+	}
+
+	resDir := ensureSubDir(contentsDir, "Resources")
+	icnsPath := filepath.Join(resDir, "icon.icns")
+
+	img, err := os.Open(p.icon)
+	if err != nil {
+		return errors.Wrapf(err, "Failed to open source image \"%s\"", p.icon)
+	}
+	defer img.Close()
+	srcImg, _, err := image.Decode(img)
+	if err != nil {
+		return errors.Wrapf(err, "Failed to decode source image")
+	}
+	dest, err := os.Create(icnsPath)
+	if err != nil {
+		return errors.Wrap(err, "Failed to open destination file")
+	}
+	defer dest.Close()
+	if err := icns.Encode(dest, srcImg); err != nil {
+		return errors.Wrap(err, "Failed to encode icns")
+	}
+
+	return nil
+}
