+++ cmd/fyne/package-unix.go
@@ -1,0 +1,81 @@
+package main
+
+import (
+	"html/template"
+	"os"
+	"os/exec"
+	"path/filepath"
+
+	"github.com/pkg/errors"
+)
+
+type unixData struct {
+	Name, Exec, Icon string
+	Local            string
+}
+
+func (p *packager) packageUNIX() error {
+	var prefixDir string
+	local := "local/"
+	tempDir := "tmp-pkg"
+
+	if p.install {
+		tempDir = ""
+	}
+
+	if _, err := os.Stat(filepath.Join("/", "usr", "local")); os.IsNotExist(err) {
+		prefixDir = ensureSubDir(ensureSubDir(p.dir, tempDir), "usr")
+		local = ""
+	} else {
+		prefixDir = ensureSubDir(ensureSubDir(ensureSubDir(p.dir, tempDir), "usr"), "local")
+	}
+
+	shareDir := ensureSubDir(prefixDir, "share")
+
+	binDir := ensureSubDir(prefixDir, "bin")
+	binName := filepath.Join(binDir, filepath.Base(p.exe))
+	err := copyExeFile(p.exe, binName)
+	if err != nil {
+		return errors.Wrap(err, "Failed to copy application binary file")
+	}
+
+	iconDir := ensureSubDir(shareDir, "pixmaps")
+	iconPath := filepath.Join(iconDir, p.name+filepath.Ext(p.icon))
+	err = copyFile(p.icon, iconPath)
+	if err != nil {
+		return errors.Wrap(err, "Failed to copy icon")
+	}
+
+	tpl, err := template.ParseFiles("templates/application.desktop", "templates/Makefile")
+	if err != nil {
+		return errors.Wrap(err, "Failed to parse templates")
+	}
+
+	appsDir := ensureSubDir(shareDir, "applications")
+	desktop := filepath.Join(appsDir, p.name+".desktop")
+	deskFile, _ := os.Create(desktop)
+
+	tplData := unixData{Name: p.name, Exec: filepath.Base(p.exe), Icon: p.name + filepath.Ext(p.icon), Local: local}
+	err = tpl.ExecuteTemplate(deskFile, "template.desktop", tplData)
+	if err != nil {
+		return errors.Wrap(err, "Failed to write desktop entry string")
+	}
+
+	if !p.install {
+		defer os.RemoveAll(filepath.Join(p.dir, tempDir))
+
+		makefile, _ := os.Create(filepath.Join(p.dir, tempDir, "Makefile"))
+		err := tpl.ExecuteTemplate(makefile, "template.makefile", tplData)
+		if err != nil {
+			return errors.Wrap(err, "Failed to write Makefile string")
+		}
+
+		tarCmd := exec.Command("tar", "zcf", p.name+".tar.gz", "-C", tempDir, "usr", "Makefile")
+		err = tarCmd.Run()
+		if err != nil {
+			return errors.Wrap(err, "Failed to create archive with tar")
+		}
+	}
+
+	return nil
+}
