+++ cmd/fyne/package-windows.go
@@ -1,0 +1,148 @@
+package main
+
+import (
+	"html/template"
+	"image"
+	"io/ioutil"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"runtime"
+	"strings"
+
+	ico "github.com/Kodeworks/golang-image-ico"
+	"github.com/josephspurrier/goversioninfo"
+	"github.com/pkg/errors"
+	"golang.org/x/mod/modfile"
+	"golang.org/x/mod/module"
+)
+
+type windowsData struct {
+	Name string
+}
+
+func (p *packager) packageWindows() error {
+	exePath := filepath.Dir(p.exe)
+
+	// convert icon
+	img, err := os.Open(p.icon)
+	if err != nil {
+		return errors.Wrap(err, "Failed to open source image")
+	}
+	defer img.Close()
+	srcImg, _, err := image.Decode(img)
+	if err != nil {
+		return errors.Wrap(err, "Failed to decode source image")
+	}
+
+	icoPath := filepath.Join(exePath, p.name+".ico")
+	file, err := os.Create(icoPath)
+	if err != nil {
+		return errors.Wrap(err, "Failed to open image file")
+	}
+
+	err = ico.Encode(file, srcImg)
+	if err != nil {
+		return errors.Wrap(err, "Failed to encode icon")
+	}
+
+	err = file.Close()
+	if err != nil {
+		return errors.Wrap(err, "Failed to close image file")
+	}
+
+	// write manifest
+	manifest := p.exe + ".manifest"
+	manifestGenerated := false
+	if _, err := os.Stat(manifest); os.IsNotExist(err) {
+		manifestGenerated = true
+		manifestFile, _ := os.Create(manifest)
+
+		tpl, err := template.ParseFiles("templates/application.manifest")
+		if err != nil {
+			return errors.Wrap(err, "Failed to parse manifest template")
+		}
+
+		tplData := windowsData{Name: p.name}
+		err = tpl.Execute(manifestFile, tplData)
+		if err != nil {
+			return errors.Wrap(err, "Failed to write manifest template")
+		}
+	}
+
+	// launch rsrc to generate the object file
+	outPath := filepath.Join(exePath, "fyne.syso")
+
+	vi := &goversioninfo.VersionInfo{}
+	vi.ProductName = p.name
+	vi.IconPath = icoPath
+	vi.ManifestPath = manifest
+
+	vi.Build()
+	vi.Walk()
+
+	arch, ok := os.LookupEnv("GOARCH")
+	if !ok {
+		arch = runtime.GOARCH
+	}
+
+	err = vi.WriteSyso(outPath, arch)
+	if err != nil {
+		return errors.Wrap(err, "Failed to write .syso file")
+	}
+
+	err = os.Remove(icoPath)
+	if err != nil {
+		return errors.Wrap(err, "Failed to remove icon")
+	} else if manifestGenerated {
+		err := os.Remove(manifest)
+		if err != nil {
+			return errors.Wrap(err, "Failed to remove manifest")
+		}
+	}
+
+	err = p.buildPackage()
+	if err != nil {
+		return errors.Wrap(err, "Failed to rebuild after adding metadata")
+	}
+
+	if p.install {
+		err := runAsAdminWindows("copy", "\"\""+p.exe+"\"\"", "\"\""+filepath.Join(os.Getenv("ProgramFiles"), p.name)+"\"\"")
+		if err != nil {
+			return errors.Wrap(err, "Failed to run as administrator")
+		}
+	}
+	return nil
+}
+
+func calculateExeName(sourceDir, os string) string {
+	exeName := filepath.Base(sourceDir)
+	/* #nosec */
+	if data, err := ioutil.ReadFile(filepath.Join(sourceDir, "go.mod")); err == nil {
+		modulePath := modfile.ModulePath(data)
+		moduleName, _, ok := module.SplitPathVersion(modulePath)
+		if ok {
+			paths := strings.Split(moduleName, "/")
+			name := paths[len(paths)-1]
+			if name != "" {
+				exeName = name
+			}
+		}
+	}
+
+	if os == "windows" {
+		exeName = exeName + ".exe"
+	}
+
+	return exeName
+}
+
+func runAsAdminWindows(args ...string) error {
+	cmd := "\"/c\""
+
+	for _, arg := range args {
+		cmd += ",\"" + arg + "\""
+	}
+
+	return exec.Command("powershell.exe", "Start-Process", "cmd.exe", "-Verb", "runAs", "-ArgumentList", cmd).Run()
+}
