+++ cmd/fyne/templates/Info.plist
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple Computer//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+<key>CFBundleName</key>
+<string>{{.Name}}</string>
+<key>CFBundleExecutable</key>
+<string>{{.ExeName}}</string>
+<key>CFBundleIdentifier</key>
+<string>{{.AppID}}</string>
+<key>CFBundleIconFile</key>
+<string>icon.icns</string>
+<key>CFBundleShortVersionString</key>
+<string>1.0</string>
+<key>NSHighResolutionCapable</key>
+<true/>
+<key>CFBundleInfoDictionaryVersion</key>
+<string>6.0</string>
+<key>CFBundlePackageType</key>
+<string>APPL</string>
+</dict>
+</plist>
