+++ cmd/fyne/templates/application.desktop
@@ -1,0 +1,5 @@
+[Desktop Entry]
+Type=Application
+Name={{.Name}}
+Exec={{.Exec}}
+Icon={{.Name}}
